Added Multiple Sampling and Improved SD Code

+++ system_node/lib/IDevice.h
@@ -27,17 +27,17 @@ class IDATA
 {
 	public:
 
-		uint8_t HW_ID	          =	  0;
+		uint8_t HW_ID				=	0;
 
-		float SYSTEM_TEMPERATURE  =   0;
+		float SYSTEM_TEMPERATURE	=	0;
 
-		float SYSTEM_HUMIDITY     =   0;
+		float SYSTEM_HUMIDITY		=	0;
 
-		float SOIL_TEMPERATURE    =   0;
+		float SOIL_TEMPERATURE		=	0;
 
-		float SOIL_MOISTURE       =   0;
+		uint8_t SOIL_MOISTURE		=	0;
 
-		float BATTERY_VOLTAGE     =   0;
+		float BATTERY_VOLTAGE		=	0;
 };
 
 #endif
